# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IntelliDesk API
  version: 0.1.0
  description: |
    与 PRD_Backend_Detailed 对齐。鉴权：请求头 `X-Tenant-Id`（必填）、`X-User-Id`（可选，演示用已存在用户 id）。
servers:
  - url: http://localhost:4000
    description: 本地后端直连（默认 PORT）
  - url: http://localhost:4001
    description: 经前端 Vite 同源代理（npm run dev:api，浏览器站点根）
...
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: X-User-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
  schemas:
...
    Ticket:
      type: object
      properties:
        id: { type: string, format: uuid }
        channelId: { type: string, description: 渠道展示名，兼容前端 }
        channelRefId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        orderId: { type: string, format: uuid, nullable: true }
        status:
          type: string
          enum: [new, todo, waiting, snoozed, resolved, spam]
        priority: { type: integer }
        sentiment: { type: string }
        intent: { type: string }
        messageProcessingStatus:
          type: string
          enum: [unread, unreplied, replied]
        subject: { type: string }
        subjectOriginal: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        firstResponseAt: { type: string, format: date-time, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        slaDueAt: { type: string, format: date-time }
        isImportant: { type: boolean }
        isFavorite: { type: boolean }
        isShared: { type: boolean }
        tags:
          type: array
          items: { type: string }
        assignedSeatId: { type: string, nullable: true }
        ownerUserId: { type: string, nullable: true }
        channelDefaultCurrency:
          { type: string, nullable: true, description: 母系统 platform-site 店铺默认货币 ISO }
        channelPlatformLanguage:
          { type: string, nullable: true, description: 母系统 platform-site 平台/站点语言 }
        motherConversationLinked:
          type: boolean
          description: 是否已关联母系统会话（externalId），本地无消息时可提示同步
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        ticketId: { type: string, format: uuid }
        senderId: { type: string }
        senderType:
          type: string
          enum: [customer, manager, agent, system, ai]
        content: { type: string }
        createdAt: { type: string, format: date-time }
        attachments:
          type: array
          items: { type: string }
        isInternal: { type: boolean }
        translatedContent: { type: string, nullable: true }
        sentPlatformText: { type: string, nullable: true }
        deliveryTargets:
          type: array
          items:
            type: string
            enum: [customer, manager]
    Customer:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        shippingCountry: { type: string }
        sentiment: { type: string }
        segments:
          type: array
          items: { type: string }
        totalOrderValue: { type: number }
        ordersCount: { type: integer }
        latestPurchaseAt: { type: string, format: date-time }
        customerSince: { type: string, format: date-time }
    LogisticsEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        subStatus: { type: string }
        description: { type: string }
        location: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        platformOrderId: { type: string }
        customerId: { type: string, format: uuid }
        channelId: { type: string }
        channelRefId: { type: string, format: uuid }
        skuList:
          type: array
          items: { type: string }
        productTitles:
          type: array
          items: { type: string }
        amount: { type: number }
        currency: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        orderStatus: { type: string }
        shippingStatus: { type: string }
        trackingNumber: { type: string, nullable: true }
        deliveryEta: { type: string, nullable: true }
        paymentStatus: { type: string }
        refundStatus: { type: string, nullable: true }
        returnStatus: { type: string, nullable: true }
        logisticsStatus: { type: string, nullable: true }
        logisticsEvents:
          type: array
          items: { $ref: '#/components/schemas/LogisticsEvent' }
        carrier: { type: string, nullable: true }
        hasVatInfo: { type: boolean }
    TicketDetail:
      allOf:
        - $ref: '#/components/schemas/Ticket'
        - type: object
          required: [messages]
          properties:
            messages:
              type: array
              items: { $ref: '#/components/schemas/Message' }
            messagesTotal:
              type: integer
              description: 工单消息总条数（可能大于本次返回的 messages 长度）
            messagesLimit:
              type: integer
              description: 本次请求应用的消息条数上限（includeMessages=0 时为 0）
            includeMessages:
              type: boolean
              description: 本次响应是否包含 messages 列表
            customer:
              allOf:
                - $ref: '#/components/schemas/Customer'
              nullable: true
            order:
              allOf:
                - $ref: '#/components/schemas/Order'
              nullable: true
    Pagination:
      type: object
      properties:
        limit: { type: integer }
        cursor: { type: string, nullable: true }
        nextCursor: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: '#/components/schemas/Ticket' }
security:
  - TenantHeader: []
paths:
...
  /api/tickets:
    get:
      tags: [tickets]
      summary: 工单列表
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: integer }
        - name: channelId
          in: query
          schema: { type: string, description: Channel UUID 或 displayName 模糊 }
        - name: messageProcessingStatus
          in: query
          schema: { type: string, enum: [unread, unreplied, replied] }
        - name: search
          in: query
          description: 匹配主题/原文主题/意图、关联订单 platformOrderId、或完整工单 UUID
          schema: { type: string }
        - name: filter
          in: query
          description: 仪表盘跳转 — sla_overdue 等
          schema:
            type: string
            enum: [unread, unreplied, replied, sla_overdue]
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pagination' }
  /api/tickets/{id}:
    get:
      tags: [tickets]
      summary: 工单详情含消息
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: messagesLimit
          in: query
          description: 返回最近若干条消息（升序），默认由 TICKET_DETAIL_MESSAGES_LIMIT 控制，最大 2000
          schema: { type: integer, minimum: 50, maximum: 2000 }
        - name: includeMessages
          in: query
          description: 设为 0 时不返回 messages（仅工单元数据），用于首包瘦身
          schema: { type: string, enum: ['0', '1'] }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketDetail' }
        '404': { description: Not found }
    patch:
      tags: [tickets]
      summary: 更新工单
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
                priority: { type: integer }
                tags: { type: array, items: { type: string } }
                messageProcessingStatus: { type: string }
                assignedSeatId: { type: string, nullable: true }
                isFavorite: { type: boolean }
                isImportant: { type: boolean }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ticket' }
  /api/tickets/{id}/messages:
    get:
      tags: [tickets]
      summary: 历史消息分页（早于 before 的若干条，时间升序返回）
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: before
          in: query
          description: ISO 时间游标，仅返回严格早于此时间的消息
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 20, maximum: 500, default: 100 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [messages, hasMore]
                properties:
                  messages:
                    type: array
                    items: { $ref: '#/components/schemas/Message' }
                  hasMore: { type: boolean }
        '400': { description: Invalid before }
        '404': { description: Not found }
    post:
      tags: [tickets]
      summary: 发送消息或内部备注
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, senderType]
              properties:
                content: { type: string }
                senderType: { type: string }
                senderId: { type: string }
                isInternal: { type: boolean, default: false }
                attachments: { type: array, items: { type: string } }
                translatedContent: { type: string }
                sentPlatformText: { type: string }
                deliveryTargets:
                  type: array
                  items: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
...